+++ MobileDataCollection.Survey/MobileDataCollection.Survey/Models/IUserAnswer.cs
@@ -9,7 +9,7 @@ namespace MobileDataCollection.Survey.Models
         /// <summary>
         /// Calculates the score of this IUserAnswer
         /// </summary>
-        /// <returns>Score without general range. Range should be defined in inheriting classes</returns>
-        int EvaluateScore();
+        /// <returns>Score without general range. Must be between 0 and 1 (inclusive)</returns>
+        float EvaluateScore();
     }
 }
